Functions rewritten to hang together in TwitterBot object.
